refactor: extract where_spec from fun_spec
Move TypeParamSpec, TypeParamKind, WhereConstraint, WhereClauseStyle,
render_type_params, emit_where_block, and emit_separate_where_block
into src/spec/where_spec.rs. These types were consumed across 11 files
but defined in fun_spec. Public API preserved via lib.rs re-exports.
