+++ tests/java/builder_functions.rs
@@ -1,10 +1,11 @@
 use sigil_stitch::code_block::CodeBlock;
 use sigil_stitch::lang::java_lang::JavaLang;
 use sigil_stitch::spec::file_spec::FileSpec;
-use sigil_stitch::spec::fun_spec::{FunSpec, TypeParamSpec};
+use sigil_stitch::spec::fun_spec::FunSpec;
 use sigil_stitch::spec::modifiers::{TypeKind, Visibility};
 use sigil_stitch::spec::parameter_spec::ParameterSpec;
 use sigil_stitch::spec::type_spec::TypeSpec;
+use sigil_stitch::spec::where_spec::TypeParamSpec;
 use sigil_stitch::type_name::TypeName;
 
 use super::golden;
@@ -32,7 +33,7 @@ fn test_function_with_doc() {
 
 #[test]
 fn test_generic_type_params_before_return_type() {
-    use sigil_stitch::spec::fun_spec::TypeParamSpec;
+    use sigil_stitch::spec::where_spec::TypeParamSpec;
 
     let tp = TypeParamSpec::new("T").with_bound(TypeName::primitive("Comparable"));
     let body = CodeBlock::of("Collections.sort(list);\nreturn list;", ()).unwrap();
