+++ tests/scala/builder_types.rs
@@ -2,10 +2,11 @@ use sigil_stitch::code_block::CodeBlock;
 use sigil_stitch::lang::scala::Scala;
 use sigil_stitch::spec::enum_variant_spec::EnumVariantSpec;
 use sigil_stitch::spec::file_spec::FileSpec;
-use sigil_stitch::spec::fun_spec::{FunSpec, TypeParamKind, TypeParamSpec};
+use sigil_stitch::spec::fun_spec::FunSpec;
 use sigil_stitch::spec::modifiers::TypeKind;
 use sigil_stitch::spec::parameter_spec::ParameterSpec;
 use sigil_stitch::spec::type_spec::TypeSpec;
+use sigil_stitch::spec::where_spec::{TypeParamKind, TypeParamSpec};
 use sigil_stitch::type_name::TypeName;
 
 use super::golden;
